// app/types/Movie.ts
export type Movie = {
  title: string;
  duration: string;
  rating: string;
  description: string;
  movie_link: string;

  writers: string[];
  directors: string[];
  stars: string[];
  countries_origin: string[];
  production_companies: string[];

  genres: string[];
  genre_token: string[];
  
  release_date: string;
  year: string;
  month: string;
  day: string;
};